Split utils between src/common and src/content-scripts/utils
